refactor: re-structure css text logic
