--- ts/display/css/background.css.ts
@@ -1,8 +1,0 @@
-export const generateBackgroundCSS = (backgroundColor: string): string =>
-  `.hexagon-wrapper {
-  background-color: ${backgroundColor};
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  }
-  `;
